Implemented CI and CD
Added GitHub Actions config for performing a CI build on normal pushes,
and a CD deployment when pushing tags. Most of this was copied from the
GitHub Actions POC project, but adapted for the multi-lib nature of the
project. Leveraged individual `gradle.properties` files to provide
library information to the Maven publish config.

Fixed a few JavaDoc warnings, and ensured JavaDoc generation is part of
the deployment.